Refactor showError, getHelperText, and useIsSubmitting functions out of useField
