+++ src/components/form/TextField.tsx
@@ -4,7 +4,8 @@ import {
 } from '@material-ui/core';
 import * as React from 'react';
 import { Except } from 'type-fest';
-import { FieldConfig, useField } from './index';
+import { FieldConfig, useField } from './useField';
+import { getHelperText, showError } from './util';
 
 export type TextFieldProps<FieldValue = string> = FieldConfig<FieldValue> & {
   name: string;
@@ -14,15 +15,11 @@ export type TextFieldProps<FieldValue = string> = FieldConfig<FieldValue> & {
 export function TextField<FieldValue = string>({
   name,
   InputProps,
-  helperText: nonErrorHelperText,
+  helperText,
   children,
   ...props
 }: TextFieldProps<FieldValue>) {
   const { input, meta, rest } = useField(name, props);
-
-  const helperText = meta.showError
-    ? meta.error || meta.submitError
-    : nonErrorHelperText;
 
   return (
     <MuiTextField
@@ -31,10 +28,8 @@ export function TextField<FieldValue = string>({
       required={props.required}
       {...rest}
       InputProps={{ ...InputProps, ...input }}
-      // always pass a truthy value, aka ' ', so layout doesn't adjust
-      // when an error is shown. This is per Material Design.
-      helperText={helperText || ' '}
-      error={meta.showError}
+      helperText={getHelperText(meta, helperText)}
+      error={showError(meta)}
     >
       {children}
     </MuiTextField>
